Fix diffusion sampling to match Protenix Algorithm 18, add 1MBN folding test
The diffusion sampler had five critical differences from Protenix:
1. Wrong sigma schedule (log-linear vs power-law with rho=7)
2. Missing predictor-corrector noise injection (gamma0=0.8, lambda=1.003)
3. Wrong initial noise scale (s_max vs sigma_data*s_max)
4. Missing step_scale_eta=1.5 on Euler steps
5. No centre-of-mass removal at each step

Also fix sample selection: rank all samples by confidence ranking_score
instead of blindly returning the first sample.

Result: CA RMSD on 1MBN improves from ~6.4A to ~1.3-4.8A (best 1.3A),
TM-score from ~0.48 to ~0.6-0.83.

Add TestFoldRealProtein integration test that folds 1MBN (sperm whale
myoglobin + heme) using Protenix weights and MSA from ColabFold server,
then asserts CA RMSD < 6A, TM-score > 0.5, lDDT > 0.5.

+++ src/helico/data.py
@@ -1169,7 +1169,7 @@ def a3m_to_msa_matrix(seqs: list[str]) -> tuple[np.ndarray, np.ndarray]:
 
     n_seqs = len(seqs)
     msa = np.full((n_seqs, L), gap_idx, dtype=np.int8)
-    deletions = np.zeros((n_seqs, L), dtype=np.int8)
+    deletions = np.zeros((n_seqs, L), dtype=np.int16)
 
     for si, seq in enumerate(seqs):
         pos = 0
